Add files via upload
Added a "Networking" folder. At this time, the "Networking" folder contains two scripts. These two scripts test TCP ports on multiple hosts and log the results to an Excel spreadsheet. One script tests a single user-defined port while the other script tests a user-defined list of multiple ports (or a single port in a list file if that's your type of thing). These scripts were written in and tested with PowerShell 7.4.4 on Windows 11. These scripts are not compatible with PowerShell 5.1 (Windows PowerShell) because they use a ForEach-Object loop with the -Parallel option for multithreading/multitasking.